Move isNotDelta
